fix rollout update
