+++ lib/commands/package/build.js
@@ -6,7 +6,7 @@ const debug = require('debug')('hyder:build');
 const { listFiles } = require('../../utils/fs');
 const hashUtil = require('../../utils/hash');
 const { log } = require('../../utils/out');
-const loadConfig = require('../../utils/loadConfig');
+const { loadConfig } = require('../../utils/product');
 
 
 module.exports = async(root, opts) => {
