+++ lib/commands/package/upload.js
@@ -35,7 +35,7 @@ module.exports = async function upload(path) {
 
   const success = await doUpload(path, { name: match[1], version: match[2] });
   if (!success) {
-    process.exit(1);
+    error('upload fail');
   }
 };
