+++ lib/commands/rollout.js
@@ -33,8 +33,9 @@ exports.list = async(name, opts) => {
 };
 
 
-exports.update = async(name, version) => {
+exports.update = async(name, opts) => {
   const url = `/admin/products/${name}/rollouts`;
+  const version = opts.V;
 
   const ok = await confirm(`Are you sure to rollout ${blue(name)}, version: ${blue(version)} ？`);
   if (!ok) {
@@ -49,10 +50,10 @@ exports.update = async(name, version) => {
     }
   });
 
-  error(red('rollout fail, response:'));
-  error(res);
-  if (!res.success) {
-    process.exit(1);
+  if (res.success) {
+    log(res);
+  } else {
+    error(res);
   }
 };
