+++ lib/index.js
@@ -1,7 +1,7 @@
 const commander = require('commander');
 const debug = require('debug')('hyder/main');
 const { log } = require('./utils/out');
-const loadConfig = require('./utils/loadConfig');
+const { loadConfig } = require('./utils/product');
 
 
 const Scope = {
@@ -142,6 +142,7 @@ function rolloutProgram(define) {
 
   program
     .command('update [product]')
+    .requiredOption('-V <version>')
     .option('--namespace <namespace>')
     .option('-y')
     .action(withCwd(define.update));
